timer: add remaining time to the component's attributes
This patch adds the amount of remaining time of a timer to the list of its state attributes. So that users can receive state changes including it.

In my specific case, it allows me to write a timer UI where I can correctly display it when paused and when the UI didn't receive the pause event (e.g. you reload the UI). Otherwise, knowing the starting date and the duration doesn't help in computing it.